minor adjustments and move into authed
