+++ src/frontend/src/routes/(new-styling)/manage/(authenticated)/security/+page.svelte
